修正无法正确设置to_idx字段的问题
