+++ code/client/pool/send.go
@@ -113,44 +113,3 @@ func (conn *Conn) SendKeepalive() {
 	case <-time.After(conn.parent.cfg.WriteTimeout):
 	}
 }
-
-// SendShellCreate send shell create message
-func (conn *Conn) SendShellCreate(id string, cfg global.Tunnel) {
-	var msg network.Msg
-	msg.To = cfg.Target
-	msg.XType = network.Msg_shell_create
-	msg.LinkId = id
-	msg.Payload = &network.Msg_Screate{
-		Screate: &network.ShellCreate{
-			Name: cfg.Name,
-			Exec: cfg.Exec,
-		},
-	}
-	select {
-	case conn.write <- &msg:
-	case <-time.After(conn.parent.cfg.WriteTimeout):
-	}
-}
-
-// SendData send shell data
-func (conn *Conn) SendShellData(to string, toIdx uint32, id string, data []byte) {
-	dup := func(data []byte) []byte {
-		ret := make([]byte, len(data))
-		copy(ret, data)
-		return ret
-	}
-	var msg network.Msg
-	msg.To = to
-	msg.ToIdx = toIdx
-	msg.XType = network.Msg_shell_data
-	msg.LinkId = id
-	msg.Payload = &network.Msg_Sdata{
-		Sdata: &network.ShellData{
-			Data: dup(data),
-		},
-	}
-	select {
-	case conn.write <- &msg:
-	case <-time.After(conn.parent.cfg.WriteTimeout):
-	}
-}
