+++ code/client/shell/forward.go
@@ -21,7 +21,16 @@ func (link *Link) remoteRead() {
 		if msg == nil {
 			return
 		}
+		link.targetIdx = msg.GetFromIdx()
 		switch msg.GetXType() {
+		case network.Msg_shell_created:
+			if msg.GetCrep().GetOk() {
+				link.OnWork <- struct{}{}
+				continue
+			}
+			logging.Error("create shell %s on tunnel %s failed, err=%s",
+				link.id, link.parent.Name, msg.GetCrep().GetMsg())
+			return
 		case network.Msg_shell_resize:
 			// TODO
 		case network.Msg_shell_data:
@@ -41,6 +50,7 @@ func (link *Link) remoteRead() {
 func (link *Link) localRead() {
 	defer utils.Recover("localRead")
 	defer link.Close()
+	<-link.OnWork
 	buf := make([]byte, 16*1024)
 	for {
 		n, err := link.stdout.Read(buf)
