+++ code/network/msg.proto
@@ -21,10 +21,11 @@ message msg {
         disconnect  = 5;
         forward     = 6;
         // shell
-        shell_create = 10;
-        shell_resize = 11;
-        shell_data   = 12;
-        shell_close  = 13;
+        shell_create  = 10;
+        shell_created = 11;
+        shell_resize  = 12;
+        shell_data    = 13;
+        shell_close   = 14;
     }
     type      _type = 1;
     string     from = 2;
@@ -39,7 +40,8 @@ message msg {
         data             _data = 13;
         // shell
         shell_create   screate = 20;
-        shell_resize   sresize = 21;
-        shell_data       sdata = 22;
+        shell_created screated = 21;
+        shell_resize   sresize = 22;
+        shell_data       sdata = 23;
     }
 }
